Refactor the output architecture.
This patch changes the output architecture in a very similar way to how
the input architecture was changed.

This simplifies things a bunch and makes adding additional encoders and
outputs simpler.

Same as the inputs refactor, this does not address the problem of having
multiple encoders per output (e.g. audio).