Only request debug layers when they're available
Previously, vulkan debug layers were always activated when compiling in
debug mode. Unfortunately, when a user compiled this library without
having a separate package installed with the debug layers, it would fail
to initialize because the lack of debug layers.

This patch only requests them if they are available on the system.