[vk-video] Get picture dimensions for download/transfer correctly
Before this change we relied on `max_coded_extent` being the video size,
wich is not always the case, especially not in mp4 files. This PR fixes
that and we extract the correct dimensions from the sps.

This reuired storing sps and pps in the decoder in addition to them
being store on the GPU.