Update to the new version of `h264_reader`.
The new version includes handling scaling lists, so this patch also
handles them correctly in our decoder. Unfortunately I'm unable to find
any h264 videos, that use scaling lists, so this is untested, but I
figure if it doesn't work somebody will sooner or later come to us with
an issue.